refactor(react-sketchapp): moved types file to same level as index and render
