+++ types/react-sketchapp/types.d.ts
